Feature: - Add selected and checked for the extra events of listbox.
*NOTE*
The call of extra events may cause a crash in VC2012RC due to compiler error.
I submitted the issue in July, but Microsoft did not close it. The ticket of the bug:753246
To avoid the crash, take the address of a function with the unary address-of operator excplicity in VC2012.